Fixed the library imports for tests on Windows platform
The `SD_SRC` directory a cygwin path and so not compatible with the
Windows Python executable.

To support running tests on Windows plaform through make.

   $ make test PYTHON=<python executable>

We need an alternative way to add the `SD_SRC` directory into the
Python system path. So we need to change the current working
directory to `SD_SRC`.